refactoring of sources
