+++ dq-core/src/main/scala/it/agilelab/bigdata/DataQuality/metrics/ColumnMetrics/BasicNumericMetrics.scala
@@ -1,7 +1,7 @@
 package it.agilelab.bigdata.DataQuality.metrics.ColumnMetrics
 
 import it.agilelab.bigdata.DataQuality.metrics.MetricCalculator
-import it.agilelab.bigdata.DataQuality.metrics.SourceProcessor.ParamMap
+import it.agilelab.bigdata.DataQuality.metrics.MetricProcessor.ParamMap
 import it.agilelab.bigdata.DataQuality.utils.{getParametrizedMetricTail, _}
 import org.isarnproject.sketches.TDigest
