+++ dq-core/src/main/scala/it/agilelab/bigdata/DataQuality/metrics/ColumnMetrics/MultiColumnMetrics.scala
@@ -1,7 +1,7 @@
 package it.agilelab.bigdata.DataQuality.metrics.ColumnMetrics
 
 import it.agilelab.bigdata.DataQuality.metrics.CalculatorStatus.CalculatorStatus
-import it.agilelab.bigdata.DataQuality.metrics.SourceProcessor.ParamMap
+import it.agilelab.bigdata.DataQuality.metrics.MetricProcessor.ParamMap
 import it.agilelab.bigdata.DataQuality.metrics.{CalculatorStatus, MetricCalculator, StatusableCalculator}
 import it.agilelab.bigdata.DataQuality.utils.{Logging, getParametrizedMetricTail, tryToString}
 import org.apache.commons.lang3.StringUtils
